+++ docs/js/examples/prevent-scroll-up.js
@@ -1,22 +1,13 @@
 new Autocomplete("preventScrollUp-output", {
   preventScrollUp: true,
 
-  onSearch: ({ currentValue }) => {
+  onSearch: async ({ currentValue }) => {
     const api = "./characters.json";
-
-    return new Promise((resolve) => {
-      fetch(api)
-        .then((response) => response.json())
-        .then((data) => {
-          const result = data
-            .filter((el) => new RegExp(currentValue, "i").test(el.name))
-            .sort((a, b) => a.name.localeCompare(b.name));
-          resolve(result);
-        })
-        .catch((error) => {
-          console.error(error);
-        });
-    });
+    const response = await fetch(api);
+    const data = await response.json();
+    return data
+      .filter((el) => new RegExp(currentValue, "i").test(el.name))
+      .sort((a, b) => a.name.localeCompare(b.name));
   },
 
   onResults: ({ matches }) => {