+++ docs/js/examples/recent-searches.js
@@ -5,22 +5,13 @@ new Autocomplete("recent-searches", {
   cache: true,
   classPreventClosing: "additional-elements",
 
-  onSearch: ({ currentValue }) => {
+  onSearch: async ({ currentValue }) => {
     const api = "./characters.json";
-
-    return new Promise((resolve) => {
-      fetch(api)
-        .then((response) => response.json())
-        .then((data) => {
-          const result = data
-            .filter((el) => new RegExp(currentValue, "i").test(el.name))
-            .sort((a, b) => a.name.localeCompare(b.name));
-          resolve(result);
-        })
-        .catch((error) => {
-          console.error(error);
-        });
-    });
+    const response = await fetch(api);
+    const data = await response.json();
+    return data
+      .filter((el) => new RegExp(currentValue, "i").test(el.name))
+      .sort((a, b) => a.name.localeCompare(b.name));
   },
 
   // this function is responsible for rendering