+++ docs/js/examples/regex.js
@@ -5,21 +5,13 @@ new Autocomplete("regex", {
   selectFirst: true,
   regex: { expression: /aaa/g, replacement: "Dr. Wong" },
 
-  onSearch: ({ currentValue }) => {
+  onSearch: async ({ currentValue }) => {
     const api = `https://rickandmortyapi.com/api/character?name=${encodeURI(
       currentValue,
     )}`;
-
-    return new Promise((resolve) => {
-      fetch(api)
-        .then((response) => response.json())
-        .then((data) => {
-          resolve(data.results);
-        })
-        .catch((error) => {
-          console.error(error);
-        });
-    });
+    const response = await fetch(api);
+    const data = await response.json();
+    return data.results;
   },
 
   onResults: ({ matches }) =>