+++ docs/js/examples/static.js
@@ -1,21 +1,12 @@
 new Autocomplete("static", {
-  onSearch: ({ currentValue }) => {
+  onSearch: async ({ currentValue }) => {
     // static file
     const api = "./characters.json";
-
-    return new Promise((resolve) => {
-      fetch(api)
-        .then((response) => response.json())
-        .then((data) => {
-          const result = data
-            .filter((el) => new RegExp(currentValue, "i").test(el.name))
-            .sort((a, b) => a.name.localeCompare(b.name));
-          resolve(result);
-        })
-        .catch((error) => {
-          console.error(error);
-        });
-    });
+    const response = await fetch(api);
+    const data = await response.json();
+    return data
+      .filter((el) => new RegExp(currentValue, "i").test(el.name))
+      .sort((a, b) => a.name.localeCompare(b.name));
   },
 
   onResults: ({ currentValue, matches }) => {